Creació API Usuaris

+++ src/main/java/com/example/backend/repo/UserRepository.java
@@ -8,6 +8,6 @@ import com.example.backend.domain.UserAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<UserAccount, Long> {
+public interface UserRepository extends JpaRepository<UserAccount, Integer> {
     Optional<UserAccount> findByEmail(String email);
 }
